Convert to TypeScript
